# Conflicts:
#	classes/class-klarna-checkout-ajax.php
#	classes/class-klarna-order.php
#	includes/checkout/create.php

+++ classes/class-klarna-to-wc.php
@@ -446,7 +446,7 @@ class WC_Gateway_Klarna_K2WC {
 			//$this->klarna_log->add( 'klarna', 'Maybe fixing order currency...' );
 		}
 
-		if ( strtoupper( $klarna_order['purchase_currency'] !== $order->get_order_currency ) ) {
+		if ( strtoupper( $klarna_order['purchase_currency'] !== $order->get_currency() ) ) {
 			if ( 'yes' === $this->klarna_debug ) {
 				//$this->klarna_log->add( 'klarna', 'Updating order currency...' );
 			}
@@ -1053,7 +1053,9 @@ class WC_Gateway_Klarna_K2WC {
 			update_user_meta( $customer_id, 'billing_first_name', $klarna_order['billing_address']['given_name'] );
 			update_user_meta( $customer_id, 'billing_last_name', $klarna_order['billing_address']['family_name'] );
 			update_user_meta( $customer_id, 'billing_address_1', $received_billing_address_1 );
-			update_user_meta( $customer_id, 'billing_address_2', $klarna_order['billing_address']['care_of'] );
+			if ( isset( $klarna_order['billing_address']['care_of'] ) ) {
+				update_user_meta( $customer_id, 'billing_address_2', $klarna_order['billing_address']['care_of'] );
+			}
 			update_user_meta( $customer_id, 'billing_postcode', $klarna_order['billing_address']['postal_code'] );
 			update_user_meta( $customer_id, 'billing_city', $klarna_order['billing_address']['city'] );
 			update_user_meta( $customer_id, 'billing_country', strtoupper( $klarna_order['billing_address']['country'] ) );
@@ -1066,7 +1068,9 @@ class WC_Gateway_Klarna_K2WC {
 				update_user_meta( $customer_id, 'shipping_first_name', $klarna_order['shipping_address']['given_name'] );
 				update_user_meta( $customer_id, 'shipping_last_name', $klarna_order['shipping_address']['family_name'] );
 				update_user_meta( $customer_id, 'shipping_address_1', $received_shipping_address_1 );
-				update_user_meta( $customer_id, 'shipping_address_2', $klarna_order['shipping_address']['care_of'] );
+				if ( isset( $klarna_order['shipping_address']['care_of'] ) ) {
+					update_user_meta( $customer_id, 'shipping_address_2', $klarna_order['shipping_address']['care_of'] );
+				}
 				update_user_meta( $customer_id, 'shipping_postcode', $klarna_order['shipping_address']['postal_code'] );
 				update_user_meta( $customer_id, 'shipping_city', $klarna_order['shipping_address']['city'] );
 				update_user_meta( $customer_id, 'shipping_country', $klarna_order['shipping_address']['country'] );
@@ -1101,12 +1105,8 @@ class WC_Gateway_Klarna_K2WC {
 				$klarna_order->acknowledge();
 				$klarna_order->fetch();
 				$klarna_order->updateMerchantReferences( array( 'merchant_reference1' => ltrim( $order->get_order_number(), '#' ) ) );
-				$order->calculate_totals( false );
-				$order->update_status( 'pending' ); // Set status to Pending Payment before completing the order.
-				$order->payment_complete( $klarna_order['klarna_reference'] );
-				$order->set_date_created( current_time( 'timestamp', true ) );
-				delete_post_meta( klarna_wc_get_order_id( $order ), '_kco_incomplete_customer_email' );
-				add_post_meta( klarna_wc_get_order_id( $order ), '_kco_payment_created', time() );
+
+				$transaction_id = $klarna_order['klarna_reference'];
 			}
 		} else { // V2 API.
 			$order->add_order_note( sprintf( __( 'Klarna Checkout payment created. Reservation number: %1$s.  Klarna order number: %2$s', 'woocommerce-gateway-klarna' ), $klarna_order['reservation'], $klarna_order['id'] ) );
@@ -1118,12 +1118,16 @@ class WC_Gateway_Klarna_K2WC {
 			$update['merchant_reference'] = array( 'orderid1' => ltrim( $order->get_order_number(), '#' ) );
 			$klarna_order->update( $update );
 
-			// Confirm local order.
+			$transaction_id = $klarna_order['reservation'];
+		}
+
+		if ( false === apply_filters( 'klarna_finalize_order_in_thank_you_page', false ) ) {
 			$order->calculate_totals( false );
 			$order->update_status( 'pending' ); // Set status to Pending Payment before completing the order.
-			$order->payment_complete( $klarna_order['reservation'] );
+			$order->payment_complete( $transaction_id );
 			$order->set_date_created( current_time( 'timestamp', true ) );
 			delete_post_meta( klarna_wc_get_order_id( $order ), '_kco_incomplete_customer_email' );
+			add_post_meta( klarna_wc_get_order_id( $order ), '_kco_payment_created', time() );
 		}
 
 		return $klarna_order;
